fix Windows build
